move corehq/{apps/hqwebapp => tabs}/templatetags/menu_tags.py
